adding R3D library, updating CMake, adding basic setup in game and ball class
